+++ Sources/PeripheryKit/Syntax/CommentCommand.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-enum CommentCommand: CustomStringConvertible, Hashable {
+public enum CommentCommand: CustomStringConvertible, Hashable {
     case ignore
     case ignoreAll
     case ignoreParameters([String])
@@ -18,7 +18,7 @@ enum CommentCommand: CustomStringConvertible, Hashable {
         return nil
     }
 
-    var description: String {
+    public var description: String {
         switch self {
         case .ignore:
             return "ignore"