+++ Tests/RetentionFixtures/testProtocolConformedByClassButNeverDirectlyUsedIsNotRetained.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-protocol FixtureProtocol54 {
-    func protocolMethod()
-}
-
-public class FixtureClass54: FixtureProtocol54 {
-    func protocolMethod() {}
-}