+++ Tests/RetentionFixtures/testRedundantProtocolThatInheritsForeignProtocol.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+protocol FixtureProtocol118: Equatable {}
+public class FixtureClass118: FixtureProtocol118 {
+    public static func == (lhs: FixtureClass118, rhs: FixtureClass118) -> Bool {
+        true
+    }
+}